+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Remove resource set concept.
+- Remove CRD management. Due to versioning issues throughout the lifecycle of
+  operators CRDs must be managed in a different way.